# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/orders/{id}":{"get":{"tags":["order-controller"],"summary":"Get order by ID","operationId":"getOrderById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Order fetched successfully","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Failed to fetch order"},"404":{"description":"Order not found","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["order-controller"],"summary":"Update order by ID","operationId":"updateOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequestDTO"}}},"required":true},"responses":{"200":{"description":"Order updated successfully","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Failed to update order"},"404":{"description":"Order not found","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Validation failed","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/products/add":{"post":{"tags":["product-controller"],"summary":"Add a new product","operationId":"addProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequestDTO"}}},"required":true},"responses":{"200":{"description":"Product added successfully","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Validation failed"},"500":{"description":"Failed to add product"}}}},"/orders/create":{"post":{"tags":["order-controller"],"summary":"Create a new order","operationId":"createOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequestDTO"}}},"required":true},"responses":{"500":{"description":"Failed to create order"},"200":{"description":"Order created successfully","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Validation failed","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/products/{id}":{"get":{"tags":["product-controller"],"summary":"Get a product by ID","operationId":"getProductById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Product not found"},"500":{"description":"Failed to fetch product"},"200":{"description":"Product fetched successfully","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["product-controller"],"summary":"Delete a product by ID","operationId":"deleteProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Product not found"},"500":{"description":"Failed to delete product"},"200":{"description":"Product deleted successfully","content":{"*/*":{"schema":{"type":"object"}}}}}},"patch":{"tags":["product-controller"],"summary":"Update a product by ID","operationId":"updateProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"description":"A map where the key is the field name and the value is the new value. Only include the fields you want to update.","content":{"application/json":{"schema":{"type":"string","example":{"description":"New product description","price":1000}}}},"required":true},"responses":{"404":{"description":"Product not found"},"200":{"description":"Product updated successfully","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Failed to update product"}}}},"/products/search":{"get":{"tags":["product-controller"],"summary":"Search for products based on name or description","operationId":"searchProducts","parameters":[{"name":"keyword","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"id"}}],"responses":{"200":{"description":"Products fetched successfully","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Failed to fetch products"},"404":{"description":"No products found"}}}},"/products/all":{"get":{"tags":["product-controller"],"summary":"Get all products","operationId":"getAllProducts","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"id"}}],"responses":{"200":{"description":"Products fetched successfully","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Failed to fetch products"},"404":{"description":"No products found"}}}},"/orders/search":{"get":{"tags":["order-controller"],"summary":"Search orders","operationId":"searchOrders","parameters":[{"name":"customerName","in":"query","required":false,"schema":{"type":"string"}},{"name":"orderId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"id"}}],"responses":{"500":{"description":"Failed to fetch orders"},"200":{"description":"Orders fetched successfully","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/orders/all":{"get":{"tags":["order-controller"],"summary":"Get all orders","operationId":"getAllOrders","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"id"}}],"responses":{"500":{"description":"Failed to fetch orders"},"200":{"description":"Orders fetched successfully","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"OrderRequestDTO":{"type":"object","properties":{"customerName":{"type":"string"},"address":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"},"status":{"type":"string"},"products":{"type":"array","items":{"$ref":"#/components/schemas/ProductRequestForOrderDTO"}}}},"ProductRequestForOrderDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"}}},"ProductRequestDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number","format":"double"},"quantity":{"type":"integer","format":"int32"}}}}}}